// Groove.SP/groove-sp-web/src/app/features/shipment/models/shipment.model.ts
export class ShipmentModel {
    id: number = null;
    shipmentNo: string = null;
    bookingNo: string = null;
    buyerCode: string = null;
    customerReferenceNo: string = null;
    modeOfTransport: string = null;
    cargoReadyDate: string = null;
    bookingDate: string = null;
    shipmentType: string = null;
    shipFrom: string = null;
    shipFromETDDate: string = null;
    shipTo: string = null;
    shipToETADate: string | null = null;
    movement: string = null;
    totalPackage: number = null;
    totalPackageUOM: string = null;
    totalUnit: number = null;
    totalUnitUOM: string = null;
    totalGrossWeight: number = null;
    totalGrossWeightUOM: string = null;
    totalNetWeight: number = null;
    totalNetWeightUOM: string = null;
    totalVolume: number = null;
    totalVolumeUOM: string = null;
    serviceType: string = null;
    incoterm: string = null;
    status: string = null;
    isFCL: boolean = null;
    isItineraryConfirmed: boolean = null;
    poFulfillmentId: number | null = null;
    executionAgentName: string = null;
    shipperName: string = null;
    consigneeName: string = null;
    lastestActivity: string = null;
    isConfirmContainer: number = null;
    isConfirmConsolidation: number = null;
}